tree depth + betty

+++ 9-binary_tree_height.c
@@ -3,17 +3,18 @@
 /**
  * binary_tree_height - a function that measures the height of a binary tree
  * @tree: is pointer to the root node of the tree to measure the height
- * @return: size_t
+ * Return: size_t
  */
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
-	int left_height = tree_height(root->left);
-    int right_height = tree_height(root->right);
-    if (left_height >= right_height)
-        return left_height + 1;
-    else
-        return right_height + 1;
+	size_t left_height = tree_height(root->left);
+	size_t right_height = tree_height(root->right);
+
+	if (left_height >= right_height)
+		return (left_height + 1);
+	else
+		return (right_height + 1);
 }
